+++ dchars/documentation/Sphinx/RSTsourcefiles/changelog.rst
@@ -7,6 +7,7 @@ CHANGELOG_TITLE
 ------------------
 
   * fixed a bug in ConfigValue.__repr__()
+  * new function : (config_ini_data.py)ConfigValuesForOneLanguage.get_configvalue_for_this_optionname()
 
 ------------------
 0.5.0 [2013_08_20]